Move parse_function into own module.
