+++ vyper/functions/utils.py
@@ -4,11 +4,11 @@ from vyper.parser.context import (
 from vyper.parser.global_context import (
     GlobalContext,
 )
-from vyper.parser.stmt import (
-    parse_body,
-)
 from vyper.parser.parser import (
     parse_to_ast,
+)
+from vyper.parser.stmt import (
+    parse_body,
 )
